requirements.txt
Flask==2.3.2
pymongo==4.4.0
python-dotenv==1.0.0
stripe==8.10.0
gunicorn==21.2.0
